단계별 적 스텟 설정 0.2
- 단계별 유닛 생성시 인덱스 초과하던 버그 수정
- HP가 인게임중 인스펙터에서 수정이 불가하게 함
- 단계별 석상 체력 구현
- 단계별 적 유닛, 드래곤 능력치 구현

+++ Assets/Script/Common/Unit/UnitGold.cs
@@ -11,7 +11,7 @@ public class UnitGold : UnitHp
     {
         Hp = (int)initHp;
         GameManager.instance.killCt++;
-        GameManager.instance.Gold += incGold[(int)PlayerPrefs.GetFloat("level") - 1];
+        GameManager.instance.Gold += incGold[PlayerPrefs.GetInt("level") - 1];
         gameObject.SetActive(false);
     }
 }
